move TIME_SYNC handler to MovementHandler since that packet is only useful in the context of player movements
